#115 moved files from Models to ModelsTest
