+++ app/src/test/java/com/example/concordia_campus_guide/ModelTests/BuildingsTest.java
@@ -1,4 +1,10 @@
-package com.example.concordia_campus_guide.Models;
+package com.example.concordia_campus_guide.ModelTests;
+
+import com.example.concordia_campus_guide.Models.Building;
+import com.example.concordia_campus_guide.Models.Buildings;
+import com.example.concordia_campus_guide.Models.Coordinates;
+import com.example.concordia_campus_guide.Models.ListOfCoordinates;
+import com.example.concordia_campus_guide.Models.Place;
 
 import org.json.JSONArray;
 import org.json.JSONException;
