+++ app/src/test/java/com/example/concordia_campus_guide/ModelTests/FloorsTest.java
@@ -1,4 +1,9 @@
-package com.example.concordia_campus_guide.Models;
+package com.example.concordia_campus_guide.ModelTests;
+
+import com.example.concordia_campus_guide.Models.Coordinates;
+import com.example.concordia_campus_guide.Models.Floor;
+import com.example.concordia_campus_guide.Models.Floors;
+import com.example.concordia_campus_guide.Models.Place;
 
 import org.junit.Before;
 import org.junit.Test;
